+++ hydra-node/src/Hydra/HeadLogic/Error.hs
@@ -59,7 +59,8 @@ data RequirementFailure tx
   | SnapshotAlreadySigned {knownSignatures :: [Party], receivedSignature :: Party}
   | AckSnNumberInvalid {requestedSn :: SnapshotNumber, lastSeenSn :: SnapshotNumber}
   | SnapshotDoesNotApply {requestedSn :: SnapshotNumber, txid :: TxIdType tx, error :: ValidationError}
-  | RecoverNotMatchingDeposit
+  | NoMatchingDeposit
+  | RequestedDepositNotActive {depositTxId :: TxIdType tx}
   deriving stock (Generic)
 
 deriving stock instance Eq (TxIdType tx) => Eq (RequirementFailure tx)