+++ hydra-node/src/Hydra/HeadLogic/State.hs
@@ -164,8 +164,9 @@ data CoordinatedHeadState tx = CoordinatedHeadState
   -- ^ Pending deposits as observed on chain. TODO: These should be actually
   -- stored outside of the 'HeadState' to allow recovery when a head is not
   -- open. See https://github.com/cardano-scaling/hydra/issues/1812
-  , currentDepositUTxO :: Maybe (UTxOType tx)
-  -- ^ Currently processed utxo to incrementally commit. Spec: Uα
+  , currentDepositTxId :: Maybe (TxIdType tx)
+  -- ^ Current/next deposit to incrementally commit. Spec: Uα
+  -- TODO: update in spec: Uα -> tx^#α
   , decommitTx :: Maybe tx
   -- ^ Pending decommit transaction. Spec: txω
   , version :: SnapshotVersion